+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/template/CriteriaDto.java
@@ -1,9 +1,11 @@
 package org.grnet.cat.dtos.template;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.EqualsAndHashCode;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
 public class CriteriaDto {
 
     @Schema(
@@ -13,6 +15,7 @@ public class CriteriaDto {
             example = "P1"
     )
     @JsonProperty("id")
+    @EqualsAndHashCode.Include
     public String id;
 
     @Schema(
@@ -22,6 +25,7 @@ public class CriteriaDto {
             example = "Principle 1"
     )
     @JsonProperty("name")
+    @EqualsAndHashCode.Include
     public String name;
 
     @Schema(
@@ -31,6 +35,7 @@ public class CriteriaDto {
             example = "The PID owner SHOULD maintain PID attributes."
     )
     @JsonProperty("description")
+    @EqualsAndHashCode.Include
     public String description;
 
     @Schema(
@@ -40,6 +45,7 @@ public class CriteriaDto {
             example = "should"
     )
     @JsonProperty("imperative")
+    @EqualsAndHashCode.Include
     public String imperative;
 
     @Schema(
@@ -49,5 +55,6 @@ public class CriteriaDto {
             example = "should"
     )
     @JsonProperty("metric")
+    @EqualsAndHashCode.Include
     public MetricDto metric;
 }